requirements.txt
beautifulsoup4==4.11.1
blis==0.7.9
bs4==0.0.1
catalogue==2.0.8
cbor==1.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
elasticsearch==7.17.6
filelock==3.8.0
fonttools==4.38.0
huggingface-hub==0.11.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
murmurhash==1.0.9
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.5.1
pathy==0.8.1
Pillow==9.3.0
preshed==3.0.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sarge==0.1.7.post1
scikit-learn==1.1.3
scipy==1.9.3
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
trec-car-tools==2.6
trectools==0.0.49
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
